Added a ``FlakyTest.retry`` combinator to be re-run tests known to be flacky
In our test suite we have a number of tests that are known to be flacky. The
``FlakyTest.retry`` combinator re-executes the same time up to N times (where
N is defaulted to 5 at the moment).

While we should of course strive to fix all sources of flackiness in tests, this
is a pragmatic solution to prevent spurious build failures.

(cherry picked from commit cfc7ee790a584350e8fc55b31dd917d9bd338e96)