Removed uneeded locking in `ScalaProject.failedCompilerInitialization`
The reason why we were using the instance locking was to avoid showing several
times the message dialog that occurs when the presentation compiler fails to
initialize.

If you look closely, you will notice that the lock is simply used to make sure
the update of the field `messageShowed` is atomic. Therefore, it is only
natural to replace the lock usage with an `AtomicBoolean` whih has very handy
operation for atomically setting its content, and returning the old value.