Never shutdown the presentation compiler in the middle of a test run
If you want to reinitialize the state of the presentation compiler,
`askRestart` should be called instead of `shutdown`. The reason is that when
calling `shutdown`, the presentation compiler dumps all queued work items, and
this can lead to lost updates.

On the other hand, for test isolation reasons, it is absolutely fine to
shutdown the presentation compiler *before*/*after* a test is executed.