integrate extract parameter refactoring
Extract Parameter adds a new parameter to a method based on an
arbitrary expression that evaluates to a value. The original expression
becomes the default value of the new parameter. Hence, the according
method calls are not modified by this refactoring.

- Allows to choose the method to add the parameter when invoked in
nested methods.
- Inline renaming of the extracted parameter

Knowns limitations:
- Does currently not support parameter extraction in methods with no
parameter list. Adding parameter lists would break compatibility with
the existing method signature and require to update all existing method
calls.